move roles to roles.ts
