+++ lib/channels.ts
@@ -97,62 +97,3 @@ export const ListChannelMembers = async (
     };
   }
 };
-
-export const CreateChannelRole = async (
-  channelId: string,
-  roleName: string,
-  optionalActions?: string[],
-  optionalMembers?: string[],
-) => {
-  const { domainId, accessToken } = await validateOrGetToken("");
-  try {
-    const role = await mgSdk.Channels.CreateChannelRole(
-      channelId,
-      roleName,
-      domainId,
-      accessToken,
-      optionalActions,
-      optionalMembers,
-    );
-    return {
-      data: role,
-      error: null,
-    };
-  } catch (err: unknown) {
-    const knownError = err as HttpError;
-    return {
-      data: null,
-      error: knownError.error || knownError.message || knownError.toString(),
-    };
-  } finally {
-    revalidatePath("/chat");
-  }
-};
-
-export const ListChannelRoles = async ({ id, queryParams }: RequestOptions) => {
-  const { domainId, accessToken } = await validateOrGetToken("");
-  try {
-    const rolesPage = await mgSdk.Channels.ListChannelRoles(
-      id as string,
-      domainId,
-      queryParams,
-      accessToken,
-    );
-    const roles = await ProcessRoles(rolesPage.roles, accessToken);
-    return {
-      data: {
-        total: rolesPage.total,
-        offset: rolesPage.offset,
-        limit: rolesPage.limit,
-        roles,
-      } as RolePage,
-      error: null,
-    };
-  } catch (err: unknown) {
-    const knownError = err as HttpError;
-    return {
-      data: null,
-      error: knownError.error || knownError.message || knownError.toString(),
-    };
-  }
-};
